# back/api/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "GPE Backend contract interface",
    "description": "This document provides an exhaustive list of all endpoints available on gpe backend.",
    "version": "0.1.1"
  },
  "servers": [
    {
      "url": "/",
      "description": ""
    },
    {
      "url": "http://localhost:8080",
      "description": "Running local dev server"
    }
  ],
...
  "paths": {
...
    "/quiz": {
      "get": {
        "operationId": "retrieveQuizzes",
        "tags": [
          "Quiz"
        ],
        "responses": {
          "200": {
            "description": "All available quizzes",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/output.quizzes.light"
                }
              }
            }
          }
        },
        "description": "Retrieve all available quizzes, quiz contains questions but not responses. Admin will see all quizzes but member role can only see quizzes for his work category and with qui status published or closed.",
        "parameters": [
          {
            "$ref": "#/components/parameters/accessToken"
          }
        ]
      },
      "post": {
        "operationId": "createNewQuiz",
        "tags": [
          "Quiz"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/accessToken"
          }
        ],
        "responses": {
          "200": {
            "description": "Quiz created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/output.quiz"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/output.message"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable entity, some validations failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/output.message"
                }
              }
            }
          }
        },
        "description": "[ADMIN] Create a new quiz entry, with work categories associated and questions. The default status is draft",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/input.quiz"
              }
            }
          },
          "required": true
        }
      }
    },
    "/quiz/stats": {
      "get": {
        "operationId": "retrieveQuizzesStats",
        "tags": [
          "Quiz"
        ],
        "responses": {
          "200": {
            "description": "All quizzes statistics",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/output.quizzes.stats"
                }
              }
            }
          }
        },
        "description": "[ADMIN] Retrieve all quizzes statistics",
        "parameters": [
          {
            "$ref": "#/components/parameters/accessToken"
          }
        ]
      }
    },
    "/quiz/{id}": {
      "get": {
        "operationId": "retrieveQuizById",
        "tags": [
          "Quiz"
        ],
        "responses": {
          "200": {
            "description": "The quiz",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/output.quiz"
                }
              }
            }
          }
        },
        "description": "Retrieve a quiz by his id, the quiz contains responses for the connected user only. If the user is an admin, he can retrieve details of all quizzes, but if the user is role member, he cannot access quizzes where his work category is not in quiz work category target.",
        "parameters": [
          {
            "$ref": "#/components/parameters/accessToken"
          },
          {
            "$ref": "#/components/parameters/id"
          }
        ]
      },
      "delete": {
        "operationId": "deleteQuizById",
        "tags": [
          "Quiz"
        ],
        "responses": {
          "204": {
            "description": "Resource successfully deleted"
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/output.message"
                }
              }
            }
          }
        },
        "description": "[ADMIN] Delete a quiz by id, this will soft delete the quiz, questions and responses are kept",
        "parameters": [
          {
            "$ref": "#/components/parameters/accessToken"
          },
          {
            "$ref": "#/components/parameters/id"
          }
        ]
      }
    },
    "/quiz/{id}/stats": {
      "get": {
        "operationId": "retrieveQuizStatsById",
        "tags": [
          "Quiz"
        ],
        "responses": {
          "200": {
            "description": "The quiz stats",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/output.quiz.stats"
                }
              }
            }
          }
        },
        "description": "[ADMIN] Retrieve quiz statistics",
        "parameters": [
          {
            "$ref": "#/components/parameters/accessToken"
          },
          {
            "$ref": "#/components/parameters/id"
          }
        ]
      }
    },
    "/quiz/{id}/response": {
      "get": {
        "operationId": "retrieveQuizByIdWithAllResponses",
        "tags": [
          "Quiz"
        ],
        "responses": {
          "200": {
            "description": "The quiz with all responses",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/output.quiz"
                }
              }
            }
          }
        },
        "description": "[ADMIN] Retrieve a quiz by his id, the quiz contains responses for all users.",
        "parameters": [
          {
            "$ref": "#/components/parameters/accessToken"
          },
          {
            "$ref": "#/components/parameters/id"
          }
        ]
      },
      "post": {
        "operationId": "answerQuiz",
        "tags": [
          "Quiz"
        ],
        "description": "Answer to a quiz, only published quizzes",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/input.quiz.responses"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "questions answered successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/output.responses"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/output.message"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable entity, some validations failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/output.message"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "$ref": "#/components/parameters/accessToken"
          },
          {
            "$ref": "#/components/parameters/id"
          }
        ]
      }
    },
    "/question/{id}/response": {
      "get": {
        "operationId": "retrieveAllResponsesOfQuestion",
        "tags": [
          "Quiz"
        ],
        "responses": {
          "200": {
            "description": "All the responses of this question",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/output.question"
                }
              }
            }
          }
        },
        "description": "[ADMIN] Retrieve response by his id, the questions contains responses for all users.",
        "parameters": [
          {
            "$ref": "#/components/parameters/accessToken"
          },
          {
            "$ref": "#/components/parameters/id"
          }
        ]
      }
    },
...
    "/quiz/{id}/publish": {
      "post": {
        "operationId": "publishQuiz",
        "tags": [
          "Quiz"
        ],
        "responses": {
          "200": {
            "description": "published quiz",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/output.message"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/output.message"
                }
              }
            }
          }
        },
        "description": "[ADMIN] Publish the quiz to be available to work categories associated",
        "parameters": [
          {
            "$ref": "#/components/parameters/accessToken"
          },
          {
            "$ref": "#/components/parameters/id"
          }
        ]
      }
    },
    "/quiz/{id}/close": {
      "post": {
        "operationId": "closeQuiz",
        "tags": [
          "Quiz"
        ],
        "responses": {
          "200": {
            "description": "quiz closed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/output.message"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/output.message"
                }
              }
            }
          }
        },
        "description": "[ADMIN] Close a quiz to not receive responses anymore",
        "parameters": [
          {
            "$ref": "#/components/parameters/accessToken"
          },
          {
            "$ref": "#/components/parameters/id"
          }
        ]
      }
    },
...
  },
  "components": {
    "parameters": {
...
      "accessToken": {
        "name": "X-Access-Token",
        "in": "header",
        "description": "The access token",
        "required": true,
        "schema": {
          "type": "string"
        }
      },
      "id": {
        "name": "id",
        "in": "path",
        "description": "The resource identifier",
        "required": true,
        "schema": {
          "type": "integer",
          "example": 1,
          "minimum": 1
        }
      },
...
    },
    "schemas": {
...
      "input.quiz": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "name of the quiz",
            "example": "quiz label"
          },
          "description": {
            "type": "string",
            "description": "description of the quiz",
            "example": "this is a description"
          },
          "workCategories": {
            "type": "array",
            "items": {
              "type": "integer",
              "description": "Work category identifier",
              "minimum": 1
            }
          },
          "questions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/input.quiz.question"
            }
          }
        },
        "required": [
          "name",
          "description",
          "workCategories",
          "questions"
        ]
      },
      "input.quiz.question": {
        "type": "object",
        "properties": {
          "description": {
            "type": "string",
            "description": "description of the new quiz question",
            "minLength": 1,
            "maxLength": 255
          },
          "questionType": {
            "type": "string",
            "enum": [
              "text",
              "single",
              "multiple"
            ],
            "example": "multiple"
          },
          "choices": {
            "description": "Provide choices if question type in not type text. Choices are displayed to the user. At least two values are required.",
            "type": "array",
            "items": {
              "type": "string",
              "description": "The available question choice",
              "example": "choice1"
            }
          }
        },
        "required": [
          "description",
          "questionType"
        ]
      },
...
      "input.quiz.response": {
        "type": "object",
        "properties": {
          "questionId": {
            "type": "integer",
            "description": "id of the question",
            "example": 1
          },
          "choices": {
            "type": "array",
            "description": "the answers chosen by the user. if the question is a multiple or single type",
            "items": {
              "type": "integer",
              "example": 1
            }
          },
          "value": {
            "type": "string",
            "description": "the answers chosen by the user. if the question is a text type",
            "example": "This is a response example"
          }
        },
        "required": [
          "questionId"
        ]
      },
      "input.quiz.responses": {
        "type": "array",
        "items": {
          "$ref": "#/components/schemas/input.quiz.response"
        }
      },
      "output.message": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "details"
          }
        },
        "required": [
          "message"
        ]
      },
...
      "output.quizzes.light": {
        "type": "array",
        "description": "Quizzes list",
        "items": {
          "$ref": "#/components/schemas/output.quiz.light"
        }
      },
      "output.quizzes.stats": {
        "type": "object",
        "properties": {
          "openQuiz": {
            "type": "integer",
            "example": 15,
            "description": "Number of open quizzes"
          },
          "avgQuestionsQuiz": {
            "type": "integer",
            "example": 12,
            "description": "Average number of questions per quiz"
          },
          "quizzes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/output.quiz.stats"
            }
          },
          "quizCreatedByYear": {
            "type": "array",
            "items": {
              "type": "object",
              "required": [
                "number",
                "year"
              ],
              "properties": {
                "year": {
                  "type": "integer",
                  "description": "The year of counter",
                  "example": 2023
                },
                "number": {
                  "type": "integer",
                  "description": "The number of quiz created",
                  "example": 15
                }
              }
            }
          }
        },
        "required": [
          "avgQuestionsQuiz",
          "openQuiz",
          "quizCreatedByYear",
          "quizzes"
        ]
      },
      "output.quiz.stats": {
        "type": "object",
        "properties": {
          "quizId": {
            "type": "integer",
            "description": "The quiz identifier",
            "example": 1
          },
          "peopleResponded": {
            "type": "integer",
            "example": 10,
            "description": "The number of people was has responded to this quiz",
            "nullable": false
          },
          "peopleConcerned": {
            "type": "integer",
            "example": 25,
            "description": "The number of people concerned by the quiz work categories"
          },
          "numberOfQuestions": {
            "type": "integer",
            "example": 3,
            "description": "The total number of questions in the quiz"
          }
        },
        "required": [
          "peopleResponded",
          "peopleConcerned",
          "numberOfQuestions",
          "quizId"
        ]
      },
      "output.quiz": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 1,
            "description": "Quiz identifier"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "name": {
            "type": "string",
            "description": "name of the quiz",
            "example": "quiz label"
          },
          "description": {
            "type": "string",
            "description": "description of the quiz",
            "example": "this is a description"
          },
          "startedAt": {
            "type": "string",
            "format": "date-time"
          },
          "closedAt": {
            "type": "string",
            "format": "date-time"
          },
          "status": {
            "type": "string",
            "enum": [
              "draft",
              "published",
              "closed"
            ],
            "default": "draft",
            "description": "status of a quiz"
          },
          "workCategories": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/output.work.category"
            }
          },
          "questions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/output.question"
            }
          },
          "creatorId": {
            "type": "integer",
            "description": "The quiz user creator",
            "example": 1
          }
        },
        "required": [
          "id",
          "createdAt",
          "updatedAt",
          "name",
          "description",
          "status",
          "workCategories",
          "questions",
          "creatorId"
        ]
      },
      "output.quiz.light": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 1,
            "description": "Quiz identifier"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "name": {
            "type": "string",
            "description": "name of the quiz",
            "example": "quiz label"
          },
          "description": {
            "type": "string",
            "description": "description of the quiz",
            "example": "this is a description"
          },
          "startedAt": {
            "type": "string",
            "format": "date-time"
          },
          "closedAt": {
            "type": "string",
            "format": "date-time"
          },
          "status": {
            "type": "string",
            "enum": [
              "draft",
              "published",
              "closed"
            ],
            "default": "draft",
            "description": "status of a quiz"
          },
          "workCategories": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/output.work.category"
            }
          },
          "creatorId": {
            "type": "integer",
            "description": "The quiz user creator",
            "example": 1
          },
          "responded": {
            "type": "boolean",
            "example": false,
            "description": "Indicates if the connected user has responded to the quiz with at least one response."
          }
        },
        "required": [
          "id",
          "createdAt",
          "updatedAt",
          "name",
          "description",
          "startedAt",
          "closedAt",
          "status",
          "workCategories",
          "creatorId",
          "responded"
        ]
      },
      "output.question": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 1,
            "description": "Question identifier"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "description": {
            "type": "string",
            "description": "description of a question"
          },
          "responses": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/output.response"
            }
          },
          "questionType": {
            "type": "string",
            "enum": [
              "text",
              "single",
              "multiple"
            ],
            "example": "single"
          },
          "choices": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "integer",
                  "description": "The choice id",
                  "example": 1
                },
                "createdAt": {
                  "type": "string",
                  "format": "date-time"
                },
                "updatedAt": {
                  "type": "string",
                  "format": "date-time"
                },
                "value": {
                  "type": "string",
                  "example": "choice1"
                },
                "type": {
                  "type": "string",
                  "example": "single",
                  "enum": [
                    "single",
                    "multiple"
                  ]
                }
              }
            }
          }
        },
        "required": [
          "id",
          "createdAt",
          "updatedAt",
          "description",
          "questionType"
        ]
      },
...
      "output.work.category": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 1,
            "description": "Work category identifier"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "name": {
            "type": "string",
            "description": "description of a question"
          }
        },
        "required": [
          "id",
          "createdAt",
          "updatedAt",
          "name"
        ]
      },
      "output.response": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 1,
            "description": "Response identifier"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "userId": {
            "type": "integer",
            "description": "The user id linked to the response",
            "example": 1
          },
          "questionId": {
            "type": "integer",
            "description": "The question id linked to the response",
            "example": 1
          },
          "responseChoiceId": {
            "type": "integer",
            "description": "The response choice in the case of response is type single or multiple",
            "example": 1
          },
          "value": {
            "type": "string",
            "description": "Response value if response if free text input",
            "example": "response type text"
          }
        },
        "required": [
          "id",
          "createdAt",
          "updatedAt",
          "userId",
          "questionId"
        ]
      },
      "output.responses": {
        "type": "array",
        "items": {
          "$ref": "#/components/schemas/output.response"
        }
      }
    }
  }
}